pyproject.toml
[tool.poetry]
name = "permutation"
version = "0.0.0"
description = "permutation toolkit for statistical learning"
authors = ["jcain <jcain@uw.edu>", "jevarts <jevarts@uw.edu>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
scikit-learn = "^1.1"
pandas = "^1.4.2"
numpy = "^1.22.4"
pandas-stubs = "^1.2.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.12.2"
tox = "^3.24.5"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
Sphinx = "^4.1.1"
mypy = "^0.941"
sphinx-rtd-theme = "^1.0.0"
sphinx_mdinclude = "^0.5.1"
jupyter = "^1.0.0"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
